bugprone-implicit-widening-of-multiplication-result

+++ src/msa.cc
@@ -183,7 +183,7 @@ auto msa(std::FILE * fp_msaout, std::FILE * fp_consout, std::FILE * fp_profile,
   auto const alnlen = find_total_alignment_length(max_insertions);
 
   /* allocate memory for profile (for consensus) and aligned seq */
-  std::vector<prof_type> profile(PROFSIZE * alnlen);
+  std::vector<prof_type> profile(static_cast<unsigned long>(PROFSIZE * alnlen));
   std::vector<char> aln_v(alnlen + 1);
   aln = aln_v.data();
   char * cons = static_cast<char *>(xmalloc(alnlen + 1));
